# docs/source/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAM3.1 Backend",
    "description": "Text-prompted segmentation over SAM3.",
    "version": "0.1.0"
  },
  "paths": {
    "/segment-from-path": {
      "post": {
        "summary": "Predict",
        "operationId": "predict_segment_from_path_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/segment-from-upload": {
      "post": {
        "summary": "Predict Upload",
        "operationId": "predict_upload_segment_from_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_predict_upload_segment_from_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_predict_upload_segment_from_upload_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          },
          "queries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Queries"
          },
          "confidence_threshold": {
            "type": "number",
            "title": "Confidence Threshold",
            "default": 0.5
          },
          "regularize": {
            "anyOf": [
              {
                "items": {
                  "type": "boolean"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regularize"
          }
        },
        "type": "object",
        "required": [
          "image",
          "queries"
        ],
        "title": "Body_predict_upload_segment_from_upload_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PredictRequest": {
        "properties": {
          "image_path": {
            "type": "string",
            "title": "Image Path"
          },
          "queries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Queries"
          },
          "output_dir": {
            "type": "string",
            "title": "Output Dir"
          },
          "confidence_threshold": {
            "type": "number",
            "title": "Confidence Threshold",
            "default": 0.5
          },
          "regularize": {
            "anyOf": [
              {
                "items": {
                  "type": "boolean"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regularize"
          }
        },
        "type": "object",
        "required": [
          "image_path",
          "queries",
          "output_dir"
        ],
        "title": "PredictRequest"
      },
      "PredictResponse": {
        "properties": {
          "image_path": {
            "type": "string",
            "title": "Image Path"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/QueryResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "image_path",
          "results"
        ],
        "title": "PredictResponse"
      },
      "PredictUploadResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/QueryResultB64"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "results"
        ],
        "title": "PredictUploadResponse"
      },
      "QueryResult": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/SegmentationObject"
            },
            "type": "array",
            "title": "Objects"
          }
        },
        "type": "object",
        "required": [
          "query",
          "objects"
        ],
        "title": "QueryResult"
      },
      "QueryResultB64": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/SegmentationObjectB64"
            },
            "type": "array",
            "title": "Objects"
          }
        },
        "type": "object",
        "required": [
          "query",
          "objects"
        ],
        "title": "QueryResultB64"
      },
      "SegmentationObject": {
        "properties": {
          "mask_path": {
            "type": "string",
            "title": "Mask Path"
          },
          "box": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Box"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "mask_path",
          "box",
          "score"
        ],
        "title": "SegmentationObject"
      },
      "SegmentationObjectB64": {
        "properties": {
          "mask_b64": {
            "type": "string",
            "title": "Mask B64"
          },
          "box": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Box"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "mask_b64",
          "box",
          "score"
        ],
        "title": "SegmentationObjectB64"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}